pyproject.toml
[project]
name = "cutlist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "plotly-express>=0.4.1",
    "seaborn>=0.13.2",
    "streamlit>=1.46.0",
    "vega-datasets>=0.9.0",
]
